refactor version detection
